tests/wpt/web-platform-tests/css/work-in-progress/opera/animations/animation-direction_alternate.html

<!DOCTYPE html>
<html>
<head>
   <title>CSS Animations - animation-direction, alternate</title>
   <link rel="author" title="Opera Software ASA" href="http://www.opera.com/">
   <link rel="help" href="https://drafts.csswg.org/css3-animations/#the-animation-direction-property-">
   <meta name="flags" content="dom">
   <meta name="assert" content="alternate is a valid animation-direction">
   <script src="/resources/testharness.js"></script>
   <script src="/resources/testharnessreport.js"></script>
   <style type="text/css">
      @keyframes slide {
         0% {
            left: 0px;
         }
         100% {
            left: 100px;
         }
      }
      div {
         width: 100px;
         height: 100px;
         background: blue;
         position: relative;
         animation-duration: 0.8s;
         animation-iteration-count: 2;
         animation-fill-mode: forwards;
         animation-direction: alternate;
      }
      .test {
         animation-name: slide;
      }
   </style>
   <script type="text/javascript">
      setup({explicit_done:true});
      onload = function() {
         var testName = document.querySelector('title').firstChild.data;
         var div = document.querySelector('div');
         var log = document.querySelectorAll('p')[1].firstChild;
         var timeout = setTimeout(function() {
            test(function(){assert_true(false);}, testName); done();
            log.data = 'FAIL';
         }, 2000);
         div.addEventListener('animationend', function() {
            clearTimeout(timeout);
            var passed = (getComputedStyle(div, null).left == '0px');
            test(function(){assert_true(passed);}, testName); done();
            log.data = (passed ? 'PASS if the animation occurred.' : 'FAIL');
         });
         div.className = 'test';
      }
   </script>
   <style>iframe{display:none}</style>
</head>
<body>
   <p>The blue square below should slide from left to right and then slide back
   to a left-aligned position.</p>
   <p>Testing...</p>
   <div></div>
</body>
</html>
